Déplacement des fonctions before* depuis le router
Pour que les tests n’importent pas le routeur.

Aussi, on se rend compte que certaines fonctions n’étaient pas utilisées…
