+++ src/scripts/questionnaire.js
@@ -1,3 +1,25 @@
+export function beforeConseils(profil, questionnaire) {
+    if (!profil.isComplete()) return questionnaire.checkPathTo('conseils', profil)
+}
+
+export function beforeSuiviIntroduction(profil, questionnaire) {
+    if (!profil.suivi_active) return questionnaire.checkPathTo('conseils', profil)
+}
+
+export function beforeDebutSymptomes(profil, questionnaire) {
+    if (!profil.suivi_active) return questionnaire.checkPathTo('conseils', profil)
+}
+
+export function beforeSuiviSymptomes(profil, questionnaire) {
+    if (!profil.symptomes_start_date) return 'debutsymptomes'
+    if (!profil.suivi_active) return questionnaire.checkPathTo('conseils', profil)
+}
+
+export function beforeSuiviHistorique(profil, questionnaire) {
+    if (!profil.symptomes_start_date) return 'debutsymptomes'
+    if (!profil.suivi_active) return questionnaire.checkPathTo('conseils', profil)
+}
+
 // Représentation de la structure du questionnaire d’orientation
 export const ORDRE = [
     'symptomesactuels',
